Move SierraRecord into the common lib
